feat: facaService Delete Tests

+++ backend/src/main/java/com/minimi/backend/facility/facamapping/service/FacaMappingServiceImpl.java
@@ -68,6 +68,9 @@ public class FacaMappingServiceImpl implements FacaMappingService {
     @Override
     @Transactional
     public void deleteFacaMapping(Long facilityId) {
+
+        if (!facaMappingRepository.existsByFaId(facilityId)) throw new NullPointerException("No Exists");
+
         facaMappingRepository.deleteAllByFaId(facilityId);
     }
